Improvements to TimeLeft estimation, by using floating point.  This goes with commit c5e0603

+++ firmware/src/Motherboard/SDCard.hh
@@ -88,7 +88,7 @@ namespace sdcard {
 
 
     /// Return the percentage of the file printed.
-    uint8_t getPercentPlayed();
+    float getPercentPlayed();
 
     /// See if there is more data available in the playback file.
     /// \return True if there is more data in the file
